% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{add_IUCN_labels}
\alias{add_IUCN_labels}
\title{Add labels from the wider IUCN database (all species)}
\usage{
add_IUCN_labels(hdr, labels, chunksize)
}
\arguments{
\item{hdr}{A base URL provided and valid API key returned by the function \link{auth_headers}}

\item{labels}{A data frame of labels to add}

\item{chunksize}{An integer specifying the chunk size for the submission}
}
\value{
A success message as a list
}
\description{
Add labels from the wider IUCN database in chunks.
}
\examples{
\dontrun{
  add_IUCN_labels(headers, labels=my_labels, chunksize=200)
}

}
\author{
Adam Varley
}
