% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{auth_headers}
\alias{auth_headers}
\title{Initiate root url and with api}
\usage{
auth_headers(api_key)
}
\arguments{
\item{api_key}{a valid API key}
}
\value{
a list containing the root url and the api key
}
\description{
Creates a base URL object that can be used as base to call enpoints
You need o get a project API key from our dashboard to use this function
}
