% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{auth_headers_dev}
\alias{auth_headers_dev}
\title{Initiate root URL with API key (Development)}
\usage{
auth_headers_dev(
  api_key,
  okala_url = "https://dev.api.dashboard.okala.io/api/"
)
}
\arguments{
\item{api_key}{A valid API key}

\item{okala_url}{The base URL for the Okala dev API (default: dev)}
}
\value{
A list containing the root URL and the API key
}
\description{
Creates a base URL object for the development Okala API. Requires a project API key.
}
\examples{
\dontrun{
  headers <- auth_headers_dev("your_api_key")
}

}
\author{
Adam Varley
}
