% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_key}
\alias{get_key}
\title{Get API key from environment variable}
\usage{
get_key()
}
\value{
The API key as a character string
}
\description{
Retrieves the API key from the environment variable OKALA_API_KEY. If the variable is not set, an error is raised.
}
\examples{
\dontrun{
  api_key <- get_key()
}

}
\author{
Adam Varley
}
