% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_media_assets}
\alias{get_media_assets}
\title{Retrieve media assets for a given project system record ID}
\usage{
get_media_assets(hdr, datatype = c("video", "audio", "image", "eDNA"), psrID)
}
\arguments{
\item{hdr}{A base URL provided and valid API key returned by the function \link{auth_headers}}

\item{datatype}{A character vector of data types c("video","audio","image","eDNA")}

\item{psrID}{Unique project system ID for which the media assets will be retrieved}
}
\value{
This function returns a tibble of media assets for the specified project system record
}
\description{
Get all of the station data associated with your project. For data types c("video","audio","image")
}
