% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_project}
\alias{get_project}
\title{Get Project Information}
\usage{
get_project(hdr = headers)
}
\arguments{
\item{hdr}{A list containing the root URL and API key, as returned by \link{auth_headers}.}
}
\value{
No return value. Displays a message indicating the active project name.
}
\description{
Retrieves information about the active project associated with the provided API key and sets it as the active project.
}
\examples{
\dontrun{
  headers <- auth_headers("your_api_key")
  get_project(headers)
}

}
\author{
Adam Varley
}
