% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_project_labels}
\alias{get_project_labels}
\title{Get project labels for either bioacoustics or camera}
\usage{
get_project_labels(
  headers,
  label_type = c("Bioacoustic", "Camera"),
  showURL = F
)
}
\arguments{
\item{headers}{containing a valid API key}

\item{label_type}{a character vector dictating the label type c('Bioacoustic','Camera')}
}
\value{
tibble of project labels
}
\description{
Labels are derived by using either suggested labels on the platform or by manulaly
adding labels from the wider database
}
