% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_station_info}
\alias{get_station_info}
\title{Get project station metadata}
\usage{
get_station_info(hdr, datatype = c("video", "audio", "image", "eDNA"))
}
\arguments{
\item{hdr}{A base URL provided and valid API key returned by the
function \link{auth_headers}}

\item{datatype}{A character vector of data types
c("video","audio","image","eDNA")}
}
\value{
An sf object containing station metadata and geometry
}
\description{
Retrieve all of the station data associated with your project, including
video, audio, image, and eDNA data types.
}
\examples{
\dontrun{
  stations <- get_station_info(headers, datatype="video")
}

}
\author{
Adam Varley
}
