% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{getIUCNLabels}
\alias{getIUCNLabels}
\title{Get labels from the wider IUCN database (all species)}
\usage{
getIUCNLabels(hdr, offset, limit, search_term = NULL)
}
\arguments{
\item{hdr}{A base URL provided and valid API key returned by the function \link{auth_headers}}

\item{offset}{An integer specifying the offset for the query}

\item{limit}{An integer specifying the limit for the query}

\item{search_term}{A character vector specifying the search term to be used (Can be left our for ful search)}
}
\value{
a list containing tabular data and pagination information for iterative calls
}
\description{
Labels are derived by using either suggested labels on the platform or by manually
adding labels from the wider database
}
