% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/okalaR-package.R
\docType{package}
\name{okalaR-package}
\alias{okalaR}
\alias{okalaR-package}
\title{okalaR: API wrapper around the okala dashboard interface}
\description{
Allows basic reading and writing of data to and from your project on the okala dashboard.
}
\author{
\strong{Maintainer}: Adam Varley \email{adam@okala.io}

}
\keyword{internal}
