% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/okalaR-package.R
\docType{package}
\name{okalaR}
\alias{okalaR-package}
\alias{okalaR}
\title{okalaR: API wrapper for the Okala dashboard}
\description{
Provides functions to interact with the Okala dashboard API, including project, station, media, and label management.
}
\author{
\strong{Maintainer}: Adam Varley \email{adam@okala.io}

}
