% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phone_observations.R
\docType{data}
\name{phone_types}
\alias{phone_types}
\title{Valid Phone Observation Types}
\format{
An object of class \code{character} of length 8.
}
\usage{
phone_types
}
\description{
Character vector of valid item types for phone observations.
}
\keyword{internal}
