% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{plot_stations}
\alias{plot_stations}
\title{Plot stations on a leaflet map}
\usage{
plot_stations(geojson_response)
}
\arguments{
\item{geojson_response}{An sf object containing station metadata and
geometry}
}
\value{
A leaflet map widget
}
\description{
Plots station locations using leaflet, with circle markers sized by
record count.
}
\examples{
\dontrun{
  plot_stations(stations)
}

}
\author{
Adam Varley
}
