% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{add_project_labels}
\alias{add_project_labels}
\title{Add project labels for either bioacoustics or camera so labeller have access to them in the Dashboard}
\usage{
add_project_labels(hdr, labeltype = c("Bioacoustic", "Camera"), labels)
}
\arguments{
\item{hdr}{A base URL provided and valid API key returned by the function \link{auth_headers}}

\item{labeltype}{A character vector specifying the label type ('Bioacoustic' or 'Camera')}

\item{labels}{A label object list like specifying the labels to be added}
}
\value{
This function returns a tibble containing project labels
}
\description{
Labels are derived by using either suggested labels on the platform or by manually
adding labels from the wider database
}
