% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_media_assets}
\alias{get_media_assets}
\title{Get media assets for a given project_system_record_id}
\usage{
get_media_assets(
  headers,
  datatype = c("video", "audio", "image"),
  project_system_record_id,
  showURL = F
)
}
\arguments{
\item{headers}{containing a valid API key}

\item{datatype}{a character vector of data types c("video","audio","image")}
}
\value{
tibble of media assets for the given project_system_record
}
\description{
Get all of the station data assocated with your project
For data types c("video","audio","image")
}
