% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_project_labels}
\alias{get_project_labels}
\title{Get project labels for either bioacoustics or camera}
\usage{
get_project_labels(hdr, labeltype = c("Bioacoustic", "Camera"))
}
\arguments{
\item{hdr}{A base URL provided and valid API key returned by the function \link{auth_headers}}

\item{labeltype}{A character vector specifying the label type ('Bioacoustic' or 'Camera')}
}
\value{
A tibble containing project labels
}
\description{
Labels are derived by using either suggested labels on the platform or by manually adding labels from the wider database.
}
\examples{
\dontrun{
  labels <- get_project_labels(headers, labeltype='Camera')
}

}
\author{
Adam Varley
}
