% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_station_info}
\alias{get_station_info}
\title{Get project station metadata}
\usage{
get_station_info(
  headers,
  datatype = c("video", "audio", "image", "eDNA"),
  showURL = F
)
}
\arguments{
\item{headers}{a valid API key}

\item{datatype}{a character vector of data types c("video","audio","image")}
}
\value{
an sf object containing station metadata
}
\description{
Get all of the station data assocated with your project
For data types c("video","audio","image","eDNA")
}
