% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{push_new_labels}
\alias{push_new_labels}
\title{Push new labels using a chunked process}
\usage{
push_new_labels(hdr, submission_records, chunksize)
}
\arguments{
\item{hdr}{A base URL provided and valid API key returned by the function \link{auth_headers}}

\item{submission_records}{A tibble containing the records to be submitted}

\item{chunksize}{An integer specifying the chunk size for the submission}
}
\value{
A list containing tabular data and pagination information for iterative calls
}
\description{
Push new labels to the platform in chunks.
}
\examples{
\dontrun{
  push_new_labels(headers, submission_records, chunksize=30)
}

}
\author{
Adam Varley
}
