% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{push_new_timestamps}
\alias{push_new_timestamps}
\title{Push new timestamps to the platform in chunks}
\usage{
push_new_timestamps(hdr, media_metadata, chunksize)
}
\arguments{
\item{hdr}{A base URL provided and valid API key returned by the function \link{auth_headers}}

\item{media_metadata}{A tibble containing the media metadata to be submitted}

\item{chunksize}{An integer specifying the chunk size for the submission}
}
\value{
A success message as a list
}
\description{
Push new timestamps to the platform in chunks.
}
\examples{
\dontrun{
  push_new_timestamps(headers, media_metadata, chunksize=100)
}

}
\author{
Adam Varley
}
