% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{sendupatedlabels}
\alias{sendupatedlabels}
\title{Send updated labels (subfunction for push_new_labels)}
\usage{
sendupatedlabels(hdr, datachunk)
}
\arguments{
\item{hdr}{A base URL provided and valid API key returned by the function \link{auth_headers}}

\item{datachunk}{A tibble containing the records to be submitted}
}
\value{
A list containing tabular data and pagination information for iterative calls
}
\description{
Submits updated labels to the platform.
}
\examples{
\dontrun{
  sendupatedlabels(headers, datachunk)
}

}
\author{
Adam Varley
}
