% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{set_segment_blank_status}
\alias{set_segment_blank_status}
\title{Set the blank status for segment labels}
\usage{
set_segment_blank_status(
  hdr,
  segment_record_ids,
  blank_status,
  chunksize = 500
)
}
\arguments{
\item{hdr}{A base URL and valid API key object returned by the function \link{auth_headers}.}

\item{segment_record_ids}{A numeric vector of segment record IDs to update.}

\item{blank_status}{A logical value (TRUE or FALSE) to set as the blank status.}

\item{chunksize}{An integer specifying the chunk size for the submission.}
}
\value{
A list containing the API response message.
}
\description{
This function updates the blank status for a given list of segment record IDs.
}
